+++ src/index.js
@@ -27,3 +27,9 @@ require('./app');
 require('./bundle');
 
 require('./../../css/openlmis.scss');
+
+if ("serviceWorker" in navigator){
+  window.addEventListener("load", () => {
+    navigator.serviceWorker.register("/sw.js");
+  })
+}